Use Euclidian algorithm to divide by two 📚
The Euclidian division algorithm allows to get the decomposition a = b * n + r,
when dividing an integer a by an integer b, where the number r (the remainder) is the
only integer satisfying 0
=< r < b. This is clearer than doing casts as they have
weird behavior in rust, which was recently changed.

Also changed the size of the icons to be multiples of four, so they are
even after dividing by two. For some reason setting icon_size to an odd
number produces blurry images.

see: https://en.wikipedia.org/wiki/Euclidean_algorithm