style: Remove bg from non-overlay scrollbar for details dialog
Sync with gnome-software, other than trough margin. Software reduces the
top margin to make it appear closer to the header bar - since the header
bar is transparent it avoids double spacing between close button and the
visible part of the scrollbar - but since we also have a non-flat header
bar there, we can't do this.